requirements.txt
absl-py==1.0.0
aiohttp==3.7.4.post0
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.2
config==0.5.1
cryptography==37.0.2
cycler==0.10.0
fake-useragent==0.1.11
filelock==3.0.12
flatbuffers==2.0
future==0.18.2
gast==0.5.3
google==3.0.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
Google-Search-API @ git+https://github.com/abenassi/Google-Search-API@546a59cc22d3260c60a2faf9afe6477168ead627
googlesearch-python==1.1.0
grpcio==1.44.0
gTTS==2.2.4
h11==0.13.0
h5py==3.6.0
helper==2.5.0
html-parser==0.2
html5lib==1.1
idna==2.10
importlib-metadata==4.11.3
joblib==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
langdetect==1.0.9
libclang==13.0.0
Markdown==3.3.6
matplotlib==3.4.3
multidict==5.1.0
nltk==3.7
numpy==1.22.3
oauthlib==3.2.0
opt-einsum==3.3.0
outcome==1.1.0
packaging==21.0
pandas==1.4.2
Pillow==9.1.0
playsound==1.3.0
ply==3.11
pocketsphinx==0.1.15
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PySocks==1.7.1
pyspellchecker==0.6.2
python-crfsuite==0.9.7
python-dateutil==2.8.2
pytz==2022.1
PyYAML==5.4.1
regex==2022.3.15
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.8
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.7.1
selenium==2.53.6
sentencepiece==0.1.91
seqeval==1.2.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0
textblob==0.15.3
tf-estimator-nightly==2.8.0.dev2021122109
tflearn==0.5.0
threadpoolctl==2.2.0
tk==0.1.0
tokenizers==0.9.3
torch==1.5.1
tqdm==4.64.0
transformers==3.5.1
trio==0.20.0
trio-websocket==0.9.2
typing-extensions==4.2.0
underthesea==1.3.2
Unidecode==1.2.0
urllib3==1.26.9
vcrpy==4.1.1
webencodings==0.5.1
Werkzeug==2.1.1
wikipedia==1.4.0
wrapt==1.14.0
wsproto==1.1.0
yarl==1.6.3
zipp==3.8.0
